fix(frontend): losing focus when creating adding a new dynamic variable
