+++ agenta-web/src/components/NewPlayground/Components/SharedEditor/types.d.ts
@@ -1,0 +1,17 @@
+import type {BaseContainerProps} from "../types"
+import type {EditorProps} from "@/components/Editor/types"
+
+export interface SharedEditorProps extends BaseContainerProps {
+    header?: React.ReactNode
+    footer?: React.ReactNode
+    editorType?: "border" | "borderless"
+    state?: "default" | "filled" | "disabled" | "readOnly" | "focus" | "typing"
+    placeholder?: string
+    handleChange: (value: string) => void
+    initialValue: any
+    editorClassName?: string
+    description?: string
+    withTooltip?: boolean
+    disabled?: boolean
+    editorProps?: EditorProps
+}
